+++ src/common/hooks/index.js
@@ -2,5 +2,6 @@ export { default as useCloseReasonOptions } from './useCloseReasonOptions';
 export * from './useLinesLimit';
 export * from './useVendor';
 export * from './useAcqMethods';
+export * from './useOrder/useOrder';
 export * from './useOrderTemplate';
 export * from './useOpenOrderSettings';