+++ src/common/utils/index.js
@@ -5,5 +5,6 @@ export * from './getAddresses';
 export * from './getAddressOptions';
 export { default as getCreateInventorySetting } from './getCreateInventorySetting';
 export * from './getUserNameById';
+export * from './getExportAccountNumbers';
 export * from './fetchExportDataByIds';
 export * from './validateDuplicateLines';