+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Disable Check for potential duplicate order lines in Settings. Refs UIOR-888.
 * Apply changes in Orders App from Plugin Find POL. Refs UIOR-900.
 * Update unopened order "Delete piece" warning message. Refs UIOR-905.
+* Validate All POLs on order that is set to export via EDIFACT have same account number. Refs UIOR-903.
 
 ## [3.0.3](https://github.com/folio-org/ui-orders/tree/v3.0.3) (2021-12-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.0.2...v3.0.3)